[Feat] 메인 왈소리 말풍선 Hidden (#22)
