+++ WAL/WAL/Screen/Main/Controller/MainViewController.swift
@@ -162,8 +162,6 @@ final class MainViewController: UIViewController {
         setupLayout()
         setupCollectionView()
         checkTime()
-        
-//        UserDefaults.standard.removeObject(forKey: "isFirst")
     }
     
     // MARK: - Init UI
@@ -298,8 +296,6 @@ extension MainViewController: UICollectionViewDelegateFlowLayout {
         }
         
         if cell.isSelected {
-            walContentView.isFirst = false
-            
             titleLabel.isHidden = false
             subTitleLabel.isHidden = false
             
@@ -417,6 +413,7 @@ extension MainViewController {
             guard let mainData = mainData else {
                 return
             }
+            print(mainData)
         }
     }
 }
